+++ src/ambition/version_details.cc
@@ -22,6 +22,7 @@ std::vector<VersionUpdates> versionDetails = {
           _("Added save/load game list scrolling with the mouse wheel or touchpad, up/down keys, pg up/pg down keys, left/right keys, and home/end keys."),
           _("Added news scrolling with the mouse wheel or touchpad, up/down keys, pg up/pg down keys, left/right keys, and home/end keys."),
           _("Added Kingdom report scrolling with the mouse wheel or touchpad, up/down keys, pg up/pg down keys, left/right keys, and home/end keys."),
+          _("Added Economy report scrolling with the mouse wheel or touchpad, up/down keys, pg up/pg down keys, left/right keys, and home/end keys."),
           _("Added inn guest list scrolling with the mouse wheel or touchpad and up/down and home/end keys."),
           _("Added Villages report scrolling with the mouse wheel or touchpad, up/down keys, pg up/pg down keys, left/right keys, and home/end keys."),
           _("Added ctrl+click priority builder request to request the nearest builder, ignoring all optimisations and what the nearest builder is doing (except constructing a new building).  Even if the builder request icon is not visible, holding Control will make it appear if there is at least one builder who could walk there."),